Add files via upload
The Inventory Control System allows a business to keep track of its inventory and profits allowing the owner to order supplies and track sales. Up to 10 different items can be tracked. Write the inventory control system with the following features:
1. The program uses an array of 10 structures. Each structure should hold the name of the item, the cost of the item, the number of items in stock, the total profit for the item and the total profit for all items.
2. The program must display the following menu. Use a switch to make the selection. Each selection should call a different function.
To choose a function, enter its letter label:
a) Show the names and number of each item in stock including the cost of each item and total value of each item in stock.
b) Show the number of units sold each time, the profit for each item in stock and the total store profit
c) Allow the owner to order more of existing items
d) Allow the owner to order new items
e) Allow the owner to enter the sale of items
f) Quit
3. Assume that the selling price is 50% greater than the latest purchase cost of an item.
4. The program must successfully execute the premise of the menu. Choices c) d) and e) require additional input..
5. Data is to be saved to a file between runs. When the program is restarted, it must first load the last status of items in stock, their cost and profit..
6. Do not use global variables. (except for the structure definition and data I/O).
7. For each item you must keep track of the item’s name, latest purchase price, number of items in stock, number of items sold, total profit from sales.